// src/types.ts
export type SymlinkCacheGroup = { [key: string]: boolean };
export interface SymlinkCache {
    dir1: SymlinkCacheGroup;
    dir2: SymlinkCacheGroup;
}

export interface CreateEntryOptions {
    entry1: Entry | undefined;
    entry2: Entry | undefined;
    state: DifferenceState;
    level: number;
    relativePath: string;
    options: Partial<SearchOptions>;
}

export type CompareResult = 0 | -1 | 1;

export interface SearchOptions {
    // TODO: improve search options as file compare takes only one option in the
    // following order compareSize -> compareDate -> compareContent

    /**
     * Compares files by size. Defaults to 'false'.
     */
    compareSize: boolean;

    /**
     *  Compares files by date of modification (stat.mtime). Defaults to 'false'.
     */
    compareDate: boolean;

    /**
     * Two files are considered to have the same date if the difference between their modification dates fits within date tolerance. Defaults to 1000 ms.
     */
    dateTolerance: number;

    /**
     *  Compares files by content. Defaults to 'false'.
     */
    compareContent: boolean;

    /**
     * Skips sub directories. Defaults to 'false'.
     */
    skipSubdirectories: boolean;

    /**
     * Ignore symbolic links. Defaults to 'false'.
     */
    skipSymlinks: boolean;

    /**
     * Ignores case when comparing names. Defaults to 'false'.
     */
    ignoreCase: boolean;

    /**
     * File name filter. Comma separated [minimatch](https://www.npmjs.com/package/minimatch) patterns.
     */
    includeFilter?: string;

    /**
     * File/directory name exclude filter. Comma separated [minimatch](https://www.npmjs.com/package/minimatch) patterns.
     */
    excludeFilter?: string;

    /**
     * File comparison handler.
     */
    compareFile: CompareFile;

    // TODO: separate specific file compare options from general
    // Only used for lineBasedFileCompare
    ignoreLineEnding: boolean;
    ignoreWhiteSpaces: boolean;
}

export interface Results {
    statistics: StatisticResults;
    differences: Difference[];
}

export interface Entry {
    name: string;
    absolutePath: string;
    path: string;
    stat?: fs.Stats;
    lstat: fs.Stats;
    isSymlink: boolean;
}
// ...
export type CompareFile = typeof defaultFileCompare | typeof lineBasedFileCompare;